fix: android additional fix
