+++ android/src/main/kotlin/matsune/video_player/PipListener.kt
@@ -12,7 +12,7 @@ class PipListener {
     private var pipRunnable: Runnable? = null
 
     @RequiresApi(Build.VERSION_CODES.N)
-    fun startPictureInPictureListenerTimer(activity: Activity, player: VideoPlayerView) {
+    fun startPictureInPictureListenerTimer(activity: Activity, player: VideoPlayer) {
         pipHandler = Handler(Looper.getMainLooper())
         pipRunnable = Runnable {
             if (isInPip != activity.isInPictureInPictureMode) {
