+++ lib/controller/video_player_value.dart
@@ -6,6 +6,7 @@ import 'package:video_player/platform/platform.dart';
 enum VideoPlayerEventType {
   initialized,
   isPlayingChanged,
+  onPlaybackStateChanged,
   positionChanged,
   bufferChanged,
   pipChanged,
@@ -21,6 +22,7 @@ class VideoPlayerValue {
   /// rest will initialize with default values when unset.
   VideoPlayerValue({
     this.eventType,
+    this.state = 0,
     this.duration,
     this.size,
     this.buffered,
@@ -35,9 +37,13 @@ class VideoPlayerValue {
     this.errorDetails,
     this.invalid,
     this.errorCode,
-  });
+    int? stateChangedTimestamp,
+  }) : stateChangedTimestamp =
+            stateChangedTimestamp ?? DateTime.now().millisecondsSinceEpoch;
 
   VideoPlayerEventType? eventType;
+
+  final int state;
 
   /// The total duration of the video.
   ///
@@ -80,10 +86,13 @@ class VideoPlayerValue {
 
   bool get isFinished => position.inSeconds == duration?.inSeconds;
 
+  int stateChangedTimestamp;
+
   /// Returns a new instance that has the same values as this current instance,
   /// except for any overrides passed in as arguments to [copyWidth].
   VideoPlayerValue copyWith({
     VideoPlayerEventType? eventType,
+    int? state,
     Duration? duration,
     Size? size,
     DurationRange? buffered,
@@ -99,8 +108,13 @@ class VideoPlayerValue {
     bool? invalid,
     int? errorCode,
   }) {
+    int? stateChangedTimestamp;
+    if (state != this.state && state == 3) {
+      stateChangedTimestamp = DateTime.now().millisecondsSinceEpoch;
+    }
     return VideoPlayerValue(
       eventType: eventType ?? this.eventType,
+      state: state ?? this.state,
       duration: duration ?? this.duration,
       size: size ?? this.size,
       buffered: buffered ?? this.buffered,
@@ -115,6 +129,8 @@ class VideoPlayerValue {
       errorDetails: errorDetails ?? this.errorDetails,
       invalid: invalid ?? this.invalid,
       errorCode: errorCode ?? this.errorCode,
+      stateChangedTimestamp:
+          stateChangedTimestamp ?? this.stateChangedTimestamp,
     );
   }
 
@@ -136,6 +152,7 @@ class VideoPlayerValue {
         'errorDescription: $errorDescription, '
         'errorDetails: $errorDetails), '
         'invalid: $invalid), '
-        'errorCode: $errorCode)';
+        'errorCode: $errorCode), '
+        'stateChangedTimestamp: $stateChangedTimestamp';
   }
 }
