+++ lib/player/video_player_with_controls.dart
@@ -1,7 +1,12 @@
+import 'dart:io';
+
+import 'package:flutter/foundation.dart';
+import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
+import 'package:flutter/services.dart';
 import 'package:video_player/controller/controller.dart';
 import 'package:video_player/controls/controls.dart';
-import 'package:video_player/platform/platform.dart';
 import 'package:video_player/subtitles/subtitles.dart';
 
 class VideoPlayerWithControls extends StatefulWidget {
@@ -78,6 +83,12 @@ class _PlayerState extends State<_Player> {
           _textureId = newTextureId;
         });
       }
+
+      if (Platform.isAndroid &&
+          widget.controller.value.eventType ==
+              VideoPlayerEventType.fullscreenChanged) {
+        widget.controller.refreshPlayer();
+      }
     };
   }
 
@@ -104,11 +115,50 @@ class _PlayerState extends State<_Player> {
 
   @override
   Widget build(BuildContext context) {
-    return _textureId == null
-        ? Container()
-        : VideoPlayerPlatform.instance.buildView(
-            _textureId!,
-            widget.controller.value.isFullscreen,
+    if (_textureId == null) {
+      return Container();
+    }
+
+    const viewType = "matsune.video_player/VideoPlayerView";
+    if (defaultTargetPlatform == TargetPlatform.iOS) {
+      return UiKitView(
+        viewType: viewType,
+        creationParamsCodec: const StandardMessageCodec(),
+        creationParams: {
+          'textureId': _textureId!,
+          'isFullscreen': widget.controller.value.isFullscreen
+        },
+      );
+    } else {
+      return ValueListenableBuilder(
+        valueListenable: widget.controller,
+        builder: (context, value, child) {
+          final androidViewType = "$viewType$_textureId";
+          return PlatformViewLink(
+            key: ValueKey(value.stateChangedTimestamp),
+            viewType: androidViewType,
+            surfaceFactory: (context, controller) {
+              return AndroidViewSurface(
+                controller: controller as AndroidViewController,
+                gestureRecognizers: const <Factory<
+                    OneSequenceGestureRecognizer>>{},
+                hitTestBehavior: PlatformViewHitTestBehavior.opaque,
+              );
+            },
+            onCreatePlatformView: (params) {
+              // NOTE: ここでcontrollerの初期化処理を行う この時textureIdを渡す
+              return PlatformViewsService.initSurfaceAndroidView(
+                id: params.id,
+                viewType: androidViewType,
+                layoutDirection: TextDirection.ltr,
+                creationParamsCodec: const StandardMessageCodec(),
+              )
+                ..addOnPlatformViewCreatedListener(params.onPlatformViewCreated)
+                ..create();
+            },
           );
+        },
+      );
+    }
   }
 }
